Improves work for #7
This will get it to work after the first unlock after a respring. Unfortunately, there is a bug that if you are in the Notification Center when you lock your device, it won't autounlock.

The setInScreenOffMode method doesn't appear to be working on iOS 13 and therefore the cause for the issue of the device only autounlocking the first time after a respring.